Generate random children
When clicking the "Bless with child" button next to a Roman, they get a
randomly generated child. The child will have the parent's clan.
Children are 50% female/male.

Currently:

* No females have a cognomen
* All males have both cognomen and agnomen
* All praenomina, cognomina, and agnomina are randomly picked from a
  generic list

The list of romans in the model was changed to a `Dict` to make it
easier to update a roman's children.